Meetup 006
This is commiting changes we made last week, ahead of today's meetup.

We wrote a `sentiment` function to use our network to check if the
sentiment of a random sentences from the internet is positive or
negative.
Looks like the higher the number (ie. closer to 1) the sentiment is more
negative. The lower the number (ie. closer to 0) the sentiment is more
positive.

We used collab and then saved file locally and converted to .py with
jupytext.